# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: LINE authenticated API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /login:
    get:
      operationId: beginBrowserLogin
      tags: [Auth]
      parameters:
        - name: returnTo
          in: query
          required: false
          schema: { type: string }
      responses:
        "302": { description: Redirect to account login }
...
  /logout:
    get:
      operationId: logoutBrowserSession
      tags: [Auth]
      parameters:
        - name: returnTo
          in: query
          required: false
          schema: { type: string }
      responses:
        "302": { description: Clear session and redirect }
...